[CortexFlow#117]: Move display events for metric tracer to helpers
